Some renaming, more tests
