+++ codegen/src/main/scala/caliban/codegen/IntrospectionClient.scala
@@ -18,7 +18,7 @@ import zio.{ IO, RIO, Task }
 
 object IntrospectionClient {
 
-  def introspect(uri: String, headers: Option[List[CommandLineArguments.Header]]): Task[Document] =
+  def introspect(uri: String, headers: Option[List[Options.Header]]): Task[Document] =
     for {
       parsedUri <- IO.fromEither(Uri.parse(uri)).mapError(cause => new Exception(s"Invalid URL: $cause"))
       baseReq   = introspection.toRequest(parsedUri)
